feat: restructure StudentFeedback component and implement feedback functionality for teachers and classrooms
